Send message to Telegram chat and test
Implement the functions to send the message to a Telegram Chat, once the message is sent also the response is parsed and checked.

The parameters to send message to the chat is taken from the yaml file configuration.

A test has been implemented sending a test message to a personal chat and checking the response.