Implement logging on file
It's now possible to configure a log file where all the program traces are saved. This can be done through the "logfile" attribute in the logging file.

In case the logging file is not set, traces are logged to Standard Output.

To implement this feature a new library has been used, called logrus. This choice has been driven by the possibility to set logging level and better manage the logging feature.